requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
python-dotenv==1.0.0
PyYAML==6.0.1
Faker==20.1.0
cryptography==41.0.7
pytest-html==4.1.1
pytest-cov==4.1.0
selenium==4.15.0
pre-commit==3.6.0
